HeadsUpStatusBarView: account for rounded corner
There has been a bug for a long time that, if
`rounded_corner_content_padding` is set to some huge value, the icon of
the heads-up notifications on the status bar will be clipped or even
completely removed from view. Setting a huge value for rounded corner
padding is required on some devices because their rounded corners are
physically huge, e.g. on some Xiaomi devices, but they are __not__
display cutouts.

The problem here is that `rounded_corner_content_padding` is not taken
into consideration in `HeadsUpStatusBarView`. `mAbsoluteStartPadding`
only considers `notification_side_paddings` and
`notification_content_margin_start`. In `onLayout`, the code tries to
compensate for cutouts and any other possible differences between the
on-screen location of the icon and `targetPadding`, which takes the
cutout insets into account. However, on a device with a value of content
padding, the `if (start != targetPadding)` is a tautology because it has
an additional padding set by a view higher in the hierarchy, which is
not accounted for in `targetPadding`. In this case, `start` can be much
larger than `targetPadding`, or even larger than `targetPadding * 2` in
my case, causing `targetPadding - start + getPaddingStart()` to evaluate
to a minus value, and as a result, the icon gets clipped because it is
now shifted in the minus direction.

This patch fixes this by simply adding the value of
`rounded_corner_content_padding` to the absolute padding value that gets
computed right at the beginning.

Change-Id: Id8e942db05734d3a6d9045fd6fb86047c1bf21e4