[WIP] sensors: hack: replace oneplus-specific light sensor type
* OnePlus specifies their own light sensor type, probably because the
sensor is under-screen and needs special logic. For now, let's just
treat it as a standard sensor and make it at least working
* This is EXTREMELY dirty. We should probably make an HAL wrapper for
this, but however, some frameworks modification would ultimately be
needed due to the fact that we need to take screen brightness AND the
screen content right above that sensor into consideration.

Change-Id: I84cc838830e59cdfef8b275925335cf7bea1f1a3