FODCircleView: move randomly to prevent burn-in
* The FOD icon is always located on the same spot in Always-on Display
mode. This poses a potential threat of burn-in since all devices with
FOD are using OLED panels.
* As a counter measure, we move the icon around just as how all the
other elements in AOD move around. We set the maximum offset to 10% of
the size of the icon so that it is within the margin of error for pretty
much every possible sensor.

Change-Id: I17204326d36e88425312f15cbc318bd01dda1e9c